+++ src/Command/DatabaseOperation.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kyzegs\DoctrineEncryptionBundle\Command;
+
+/** @internal */
+enum DatabaseOperation: string
+{
+    case ENCRYPT = 'encrypt';
+    case DECRYPT = 'decrypt';
+    case ROTATE = 'rotate';
+}